My solution for p1
Hey guys! I tried to solve p1 so that I was thorough with the program for the quiz. I think everything works fine. Since this wasn't of any official use but just for satisfaction, I didn't bother commenting it in too much detail but still commenting is reasonably done. I reckon my implementation of GradeEstimator is a little different than Matt's. And lastly, while working on this solution, I realized I had made a mistake on ScoreIterator. Since the category of the score was a string but only of a single letter, I accidentally used == to compare it thinking it to be a character and not .equals() method. I apologize to Matt and everyone else who faced a problem because of that.